+++ source/ChanSort/MainForm.cs
@@ -2157,8 +2157,16 @@ namespace ChanSort.Ui
       this.bar1.Visible = false;
       GlobalImageCollection.Scale(absScaleFactor.Height, false);
       var newSize = this.ClientSize;
-      if (this.adjustWindowLocationOnScale) // adjust WindowStartPosition "CenterScreen" to new window size
-        this.Location = new Point(this.Left - (newSize.Width - oldSize.Width) / 2, this.Top - (newSize.Height - oldSize.Height) / 2);
+      if (this.adjustWindowLocationOnScale) // adjust WindowStartPosition "CenterScreen" to new window size, but don't move upper left corner off-screen
+      {
+        var screen = Screen.FromPoint(new Point(this.Left + this.Width / 2, this.Top + this.Height / 2));
+        this.Bounds = new Rectangle(
+          Math.Max(screen.Bounds.Left, this.Left - (newSize.Width - oldSize.Width) / 2), 
+          Math.Max(screen.Bounds.Top, this.Top - (newSize.Height - oldSize.Height) / 2),
+          Math.Min(screen.Bounds.Width, this.Width),
+          Math.Min(screen.Bounds.Height, this.Height));
+      }
+
       this.bar1.Visible = true;
       this.ResumeRedraw();
     }